+++ 3rdParty/s2/geometry/s2regionintersection.cc
@@ -13,7 +13,7 @@ S2RegionIntersection::S2RegionIntersection(vector<S2Region*>* regions) {
 }
 
 S2RegionIntersection::~S2RegionIntersection() {
-  for (int i = 0; i < regions_.size(); ++i) {
+  for (size_t i = 0; i < regions_.size(); ++i) {
     delete regions_[i];
   }
   regions_.clear();
