+++ 3rdParty/s2/geometry/s2regionunion.cc
@@ -13,7 +13,7 @@ S2RegionUnion::S2RegionUnion(vector<S2Region*>* regions) {
 }
 
 S2RegionUnion::~S2RegionUnion() {
-  for (int i = 0; i < regions_.size(); ++i) {
+  for (size_t i = 0; i < regions_.size(); ++i) {
     delete regions_[i];
   }
   regions_.clear();
